Modif mineure du tester sh

+++ test.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
-rm sortie_test
+if [ -e sortie_test ]; then rm sortie_test; fi
+echo "###############################"
 echo "-parse-only sur docs/tests/syntax/bad/*"
 for file in docs/tests/syntax/bad/*.cpp
 do
 ./minic++ -parse-only $file
 done
 
-echo "\n \n -parse-only sur le reste"
-for file in docs/tests/syntax/good/*.cpp docs/tests/typing/*/*.cpp docs/tests/exec/*.cpp
+echo
+echo "###############################"
+echo " -parse-only sur le reste"
+for file in docs/tests/syntax/good/*.cpp 
 do
 ./minic++ -parse-only $file
 done
